+++ FBSDKCoreKit/FBSDKCoreKitTests/Internal/Helpers/TestAppEventsStateProvider.swift
@@ -23,7 +23,7 @@ class TestAppEventsStateProvider: NSObject, AppEventsStateProviding {
   var capturedAppID: String?
   var isCreateStateCalled = false
 
-  func createState(tokenString: String?, appID: String?) -> AppEventsState {
+  func createState(tokenString: String, appID: String) -> AppEventsState {
     isCreateStateCalled = true
     capturedTokenString = tokenString
     capturedAppID = appID